requirements.txt
-i https://pypi.org/simple
asgiref==3.8.1; python_version >= '3.8'
django==5.0.6; python_version >= '3.10'
joblib==1.4.2; python_version >= '3.8'
numpy==1.26.4; python_version >= '3.9'
pandas==2.2.2; python_version >= '3.9'
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pytz==2024.1
scikit-learn==1.4.2; python_version >= '3.9'
scipy==1.13.0; python_version >= '3.9'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sklearn.preprocessing==0.1.0
sqlparse==0.5.0; python_version >= '3.8'
threadpoolctl==3.5.0; python_version >= '3.8'
tzdata==2024.1; python_version >= '2'
